requirements.txt
dash
dash-table
pandas
plotly
altair
streamlit
gunicorn
snowflake-connector-python
dotenv
